Plugins: categorize linux.tty_check as a malware plugin
